less arguments for taskFactoryBuilder but more responsibilities
